+++ seogi/BJ/BJ_2961.java
@@ -1,0 +1,24 @@
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.util.StringTokenizer;
+
+/*
+ * https://www.acmicpc.net/source/79776832
+ * java8 / 131876 KB / 760 ms /
+ */
+public class Main {
+  public static void main(String[] args) throws IOException {
+    BufferedReader bf = new BufferedReader(new InputStreamReader(System.in));
+    BufferedWriter bw=  new BufferedWriter(new OutputStreamWriter(System.out));
+    StringTokenizer st = new StringTokenizer(bf.readLine());
+    StringBuilder sb = new StringBuilder();
+    int N = Integer.parseInt(st.nextToken());
+    int arr[][] = new int[N][2];
+    for (int i = 0; i < N; i++) {
+
+    }
+  }
+}
